Change some arguments from unsigned int to size_t.
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@2229 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qm/include/qmmessageoperation.h
@@ -29,8 +29,8 @@ public:
 
 public:
 	virtual bool isCanceled() = 0;
-	virtual void setCount(unsigned int nCount) = 0;
-	virtual void step(unsigned int nStep) = 0;
+	virtual void setCount(size_t nCount) = 0;
+	virtual void step(size_t nStep) = 0;
 	virtual void show() = 0;
 };
