+++ q3/qm/src/model/messagestore.h
@@ -213,8 +213,8 @@ public:
 	virtual ~MessageStoreSalvageCallback();
 
 public:
-	virtual void setCount(unsigned int nCount) = 0;
-	virtual void step(unsigned int nStep) = 0;
+	virtual void setCount(size_t nCount) = 0;
+	virtual void step(size_t nStep) = 0;
 	virtual bool salvage(const Message& msg) = 0;
 };
